pyproject.toml
[tool.poetry]
name = "repl_python3_beta_firebolt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com "]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.7.3"
aiohttp="^3.6.2"
async-timeout ="^3.0.1"
attrs ="^19.3.0"
beautifulsoup4 ="^4.9.1"
bs4 ="^0.0.1"
certifi ="^2020.6.20"
chardet ="^3.0.4"
idna ="^2.10"
lavalink ="^3.1.2"
multidict ="^4.7.6"
mysql-connector ="^2.2.9"
psutil ="^5.7.2"
python-dotenv ="^0.14.0"
requests ="^2.24.0"
soupsieve ="^2.0.1"
typing-extensions ="^3.7.4.2"
urllib3 ="^1.25.10"
websockets ="^8.1"
yarl ="^1.5.1"
mysql-connector-python ="^8.0.21"
Flask = "^2.0.1"
fortune="^1.1.0"
neko = "^0.0.11"
cmds = "^0.1.1"
editdistance = "^0.5.3"
utils = "^1.0.1"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core =1.0.0"]
build-backend = "poetry.core.masonry.api"
